kernel/int: consolidate interrupt vector number to `int32`
This is no-op for 32 bit platforms because `int32` is defined as `long` there.

Change interrupt vector number from 64 bits to 32 bits for 64 bit platforms.

+++ headers/os/drivers/dm2/device/InterruptController.h
@@ -7,7 +7,7 @@ class InterruptControllerDeviceFdt {
 public:
 	static inline const char ifaceName[] = "interrupt_controller/fdt";
 
-	virtual status_t GetVector(const uint32* intrData, uint32 intrCells, long* vector) = 0;
+	virtual status_t GetVector(const uint32* intrData, uint32 intrCells, int32* vector) = 0;
 
 protected:
 	~InterruptControllerDeviceFdt() = default;
