+++ headers/private/kernel/arch/generic/generic_int.h
@@ -12,10 +12,10 @@
 
 class InterruptSource {
 public:
-	virtual void EnableIoInterrupt(int irq) = 0;
-	virtual void DisableIoInterrupt(int irq) = 0;
-	virtual void ConfigureIoInterrupt(int irq, uint32 config) = 0;
-	virtual void EndOfInterrupt(int irq) = 0;
+	virtual void EnableIoInterrupt(int32 irq) = 0;
+	virtual void DisableIoInterrupt(int32 irq) = 0;
+	virtual void ConfigureIoInterrupt(int32 irq, uint32 config) = 0;
+	virtual void EndOfInterrupt(int32 irq) = 0;
 	virtual int32 AssignToCpu(int32 irq, int32 cpu) = 0;
 };
 
@@ -24,11 +24,11 @@ extern "C" {
 
 // `_ex` functions must be used if using `generic_int`
 
-status_t reserve_io_interrupt_vectors_ex(long count, long startVector,
+status_t reserve_io_interrupt_vectors_ex(int32 count, int32 startVector,
 	enum interrupt_type type, InterruptSource* source);
-status_t allocate_io_interrupt_vectors_ex(long count, long *startVector,
+status_t allocate_io_interrupt_vectors_ex(int32 count, int32 *startVector,
 	enum interrupt_type type, InterruptSource* source);
-void free_io_interrupt_vectors_ex(long count, long startVector);
+void free_io_interrupt_vectors_ex(int32 count, int32 startVector);
 
 }
